ci: fix 'saveable' test-case
The following commit fixes an issue with the `saveable.spec.js` api
test, where due to a quick `ADDED/DELETE` event, NSFW was not properly
notified of the event, causing the test to fail when attempting to
verify its assertions. The fix includes waiting a bit before deleting
the resource so NSFW can properly listen to both `ADDED` and
`DELETED` events.

Signed-off-by: vince-fugnitto <vincent.fugnitto@ericsson.com>